==> 03_Content_Embedding-Cold_Start.ipynb <==
# Content Embedding as a Solution to the Cold Start Problem

To tackle the cold start problem in recommender systems, one effective approach is to utilize content embedding techniques. By embedding the content of movies into a high-dimensional vector space, we can capture meaningful representations of their features. These embeddings enable us to quantify similarities between movies based on their content characteristics, facilitating accurate recommendations even for new or less-known items.

In the following sections, we will explore three different methods for embedding movie content. Each method offers unique advantages and considerations, providing flexibility in addressing various aspects of the cold start problem. Through these techniques, we aim to enhance the performance and robustness of our recommender system, ensuring personalized and relevant recommendations for users across diverse scenarios.

from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.metrics.pairwise import cosine_similarity
import pandas as pd

df = pd.read_csv('./data/movies.csv')
df

dataset = pd.DataFrame()
dataset['movieId']= df['movieId']
dataset['title']= df['title']
dataset['year']= df['year']
dataset['content'] = df['title'] +' (' + df['year'].astype(str) + ')' + '; ' + df['cast_and_crew'] + '; ' + df['genre']
df['content'] = df['title'] +' (' + df['year'].astype(str) + ')' + '; ' + df['cast_and_crew'] + '; ' + df['genre']
# df['content'] = df['title'] +' (' + df['year'] + ')' + '; ' + df['cast_and_crew']

dataset

dataset

## 1. Cosine similarity

Let's generate recommendations for a given movie based on content similarity. It utilizes a count matrix and cosine similarity to calculate the similarity scores between movies. If the input movie is found in the dataset, the function returns a list of the top 20 recommended movies based on content similarity.

def get_recommendations(movie):
        # Creating a count Matrix
    cv = CountVectorizer()
    count_matrix = cv.fit_transform(dataset['content'])

    # Creating a similarity score matrix
    similarity = cosine_similarity(count_matrix)
    
    if movie in dataset['title'].unique():
        idx = dataset.loc[dataset['title']==movie].index[0]
        lst = list(enumerate(similarity[idx]))
        lst = sorted(lst, key= lambda x:x[1], reverse=True)
        lst = lst[1:20]

        movie_list = []
        for i in range(len(lst)):
            idx_sim = lst[i][0]
            movie_list.append(dataset['title'][idx_sim]+' ('+dataset['year'][idx_sim].astype(str)+')')

    return movie_list

get_recommendations('Jaws')

 In the case of the 'Jaws' movie, while the recommender system correctly suggests sequels and movies of similar genres, it also includes unrelated comedy films like 'Funny Bones' and 'B.S. I Love You'.
 
 Let's see if we can improve our recommender system.

### Explore Using Stopwords

from nltk.corpus import stopwords
import nltk
nltk.download('stopwords')

def get_recommendations(movie):
    # Creating a count Matrix
    stop_words = list(stopwords.words('english'))
    cv = CountVectorizer(stop_words=stop_words)
    count_matrix = cv.fit_transform(dataset['content'])

    # Creating a similarity score matrix
    similarity = cosine_similarity(count_matrix)
    
    if movie in dataset['title'].unique():
        idx = dataset.loc[dataset['title']==movie].index[0]
        lst = list(enumerate(similarity[idx]))
        lst = sorted(lst, key= lambda x:x[1], reverse=True)
        lst = lst[1:20]

        movie_list = []
        for i in range(len(lst)):
            idx_sim = lst[i][0]
            movie_list.append(dataset['title'][idx_sim]+' ('+dataset['year'][idx_sim].astype(str)+')')

    return movie_list

get_recommendations('Jaws')

It doesn't get any better.

## Utilizing TF-IDF for Improved Recommendations

Incorporating TF-IDF (Term Frequency-Inverse Document Frequency) can significantly enhance the quality of recommendations by providing a more nuanced representation of movie content. TF-IDF assigns weights to words based on their frequency in a document (movie) relative to the entire dataset, emphasizing terms that are unique and informative.

def get_recommendations(movie):
    # Creating a TF-IDF Matrix
    tfidf_vectorizer = TfidfVectorizer(stop_words='english')
    tfidf_matrix = tfidf_vectorizer.fit_transform(dataset['content'])

    # Creating a similarity score matrix
    similarity = cosine_similarity(tfidf_matrix)

    if movie in dataset['title'].unique():
        idx = dataset.loc[dataset['title'] == movie].index[0]
        lst = list(enumerate(similarity[idx]))
        lst = sorted(lst, key=lambda x: x[1], reverse=True)
        lst = lst[1:20]

        movie_list = []
        for i in range(len(lst)):
            idx_sim = lst[i][0]
            movie_list.append(dataset['title'][idx_sim] + ' (' + str(dataset['year'][idx_sim]) + ')')

    return movie_list

get_recommendations('Jaws')

Utilizing TF-IDF indeed offers several advantages, especially in capturing the significance of directors' names, actors, and genres within movie descriptions. By incorporating TF-IDF, the recommendation system can effectively highlight these important features and provide more accurate and relevant recommendations to users.

## Recommender System Using Embedding Databases

Embedding databases offer a powerful approach to enhancing recommender systems by providing rich representations of items such as movies, products, or documents. These databases store embeddings, which are dense, high-dimensional vectors that capture semantic relationships between items based on their features or content. By leveraging embedding databases, we can build recommender systems that offer personalized and relevant recommendations to users.

**Benefits of Using Embedding Databases**:
- **Rich Representation**: Embedding databases capture complex relationships between items, enabling a comprehensive understanding of their features or content.
- **Semantic Understanding**: By leveraging embeddings, recommender systems can interpret and capture the semantic meaning of items, leading to more accurate recommendations.
- **Flexibility**: Embedding databases offer flexibility in storing and accessing embeddings for diverse types of items, making them suitable for building recommender systems for various domains.
- **Scalability**: Embedding databases can handle large-scale datasets efficiently, allowing for the storage and retrieval of embeddings for millions or even billions of items.

By leveraging embedding databases, we can develop robust and effective recommender systems that provide personalized and relevant recommendations to users, enhancing their overall experience and satisfaction. Embedding databases offer a versatile and scalable solution for building recommender systems across diverse domains and applications.

# Uncomment to install chromadb properly
# !python -m pip install -U sentence-transformers chromadb

import pandas as pd

df = pd.read_csv("./movies/movies.csv")
df['content'] = 

import chromadb
from chromadb.utils import embedding_functions

import chromadb
from chromadb.utils import embedding_functions

CHROMA_DATA_PATH = "./movies/chroma"
EMBED_MODEL = "all-MiniLM-L6-v2"
COLLECTION_NAME = "movies"

client = chromadb.PersistentClient(path=CHROMA_DATA_PATH)

embedding_func = embedding_functions.SentenceTransformerEmbeddingFunction(model_name=EMBED_MODEL)

collection = client.create_collection(
    name=COLLECTION_NAME,
    embedding_function=embedding_func,
    metadata={"hnsw:space": "cosine"},
    get_or_create=True
)

# Uncomment in case you have created a collection then delete it
# client.delete_collection(COLLECTION_NAME)

documents = [row['content'] for _, row in df.iterrows()]
ids = [f"{row['movieId']}" for _, row in df.iterrows()]
genres = [{"genre": g} for g in df['genre']]

len(df), len(documents), len(ids), len(genres)

collection.add(
    ids=ids,
    documents=documents,
    metadatas=genres,
)

# Retrieve movies
movie = 'Jaws'
query_results = collection.query(
    query_texts=[f"Find me some movies related to {movie}"],
    n_results=20,
)

query_results["documents"]

# Retrieve documents related to Jaws by genre
query_results = collection.query(
    query_texts=[f"Find me some movies related to {movie}"],
    where={"genre": {"$in": ["Horror", "Thriller"]}},
    n_results=20,
)

query_results["documents"]
