==> 04_Build_and_Evaluate_Models.ipynb <==
# Colaborative Filtering Recommender System


## Surprise

With surprise Library, we will benchmark the following algorithms. We use "rmse" as our accuracy metric for the predictions

import pandas as pd
from surprise import Dataset
from surprise import Reader
from surprise import accuracy
from surprise.model_selection import cross_validate
from surprise.model_selection import train_test_split
from surprise import SVD, SVDpp, SlopeOne, NMF, NormalPredictor, KNNBaseline, \
    KNNBasic, KNNWithMeans, KNNWithZScore, BaselineOnly, CoClustering


### Preparing the dataset

ratings_df = pd.read_csv('./data/ratings_cleaned.csv')
ratings_df

movies_df = pd.read_csv('./data/movies.csv')
movies_df['title_year'] = movies_df['title'] + ' (' + movies_df['year'].astype(str) + ')'
movies_df

reader = Reader(line_format='user item rating', rating_scale=(0.5,10))

data = Dataset.load_from_df(df=ratings_df[['userId', 'movieId', 'rating']], reader=reader)

### Benchmarking

algo_list = [SVD(), SVDpp(), SlopeOne(), NMF(), NormalPredictor(), KNNBaseline(), KNNBasic(), KNNWithMeans(), KNNWithZScore(), BaselineOnly(), CoClustering()]
benchmark = []

for algo in algo_list:
    results = cross_validate(algo=algo, data=data, measures=["rmse"], cv=3, n_jobs=-1, verbose=False)
    algo_name = str(algo).split(' ')[0].split('.')[-1]
    print(algo_name)
    tmp = pd.DataFrame.from_dict(results).mean(axis=0)
    tmp['Algorithm'] = algo_name
    benchmark.append(tmp)

pd.DataFrame(benchmark).set_index('Algorithm').sort_values('test_rmse')

## Training and Prediction

Since the `SVDpp` algorithm yielded the best RMSE, we will utilize it for training and prediction using a simple grid search.

from surprise.model_selection import GridSearchCV

param_grid = {'n_factors': [10, 15, 20, 25, 30], 
              'n_epochs': [25, 30, 35, 40], 
              'lr_all': [0.02, 0.03],
              'reg_all': [0.01, 0.02, 0.03],
              'random_state': [0]}

gs = GridSearchCV(algo_class=SVDpp, param_grid=param_grid, measures=['rmse'], cv=3, n_jobs=-1 )

gs.fit(data)

# best RMSE score
print(gs.best_score['rmse'])

# combination of parameters that gave the best RMSE score
print(gs.best_params['rmse'])

# 2.0321557585693815
# {'n_factors': 30, 'n_epochs': 40, 'lr_all': 0.03, 'reg_all': 0.02, 'random_state': 0}

You can now use these optimal hyperparameters to train your SVDpp model on the entire dataset and make predictions.

trainset = data.build_full_trainset()

best_parameters = {'n_factors': 30, 'n_epochs': 40, 'lr_all': 0.03, 'reg_all': 0.02, 'random_state': 0}
# best_parameters = gs.best_params['rmse']
# Create an SVD algorithm with the best hyperparameters
optimal_svd = SVDpp(**best_parameters)

# Train the algorithm on the training set
optimal_svd.fit(trainset)

trainset.all_items()

USER_ID = 463 # raw
print(f"RAW USER: {USER_ID}")
inner_uid = optimal_svd.trainset.to_inner_uid(USER_ID) # inner
print(f"INNER USER: {inner_uid}")
movies_rated_inner = [int(item[0]) for item in optimal_svd.trainset.ur[inner_uid]] # inner
movies_rated = [int(optimal_svd.trainset.to_raw_iid(item)) for item in movies_rated_inner] # inner -> raw
list_rated = movies_df[movies_df['movieId'].isin(movies_rated)]['title'].tolist() # raw

print()
print("RATED:")
for e in list_rated[:10]:
    print(e)

unrated_movies_inner = [id for id in optimal_svd.trainset.all_items() if id not in movies_rated_inner] # inner
predictions = [optimal_svd.predict(uid=USER_ID, iid=optimal_svd.trainset.to_raw_iid(item)) for item in unrated_movies_inner] # raw -> raw

# Sort predictions by estimated rating in descending order
sorted_predictions = sorted(predictions, key=lambda x: x.est, reverse=True) # raw

# Predictions for USER
top_n = []
for _, iid, _, est, _ in sorted_predictions[:20]:
    top_n.append(int(iid))

list_movies = movies_df[movies_df['movieId'].isin(top_n)]['title'].tolist()

print()
print("RECOMMENDED:")
for e in list_movies:
    print(e)

This result is not exceptional, likely due to the highly sparse nature of the dataset. However, what could we do next? An immediate solution is to explore content-based filtering instead of collaborative filtering.

## Content-Base Filtering

**Under Development**
